+++ src/components/common/Nav.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import "./Nav.css";
+import "./styles/Nav.css";
 
 export default function Nav() {
   return (
